rename finetune_albert_job.py to match with inference job.
